Reconfigure remote.origin.fetch when switching branches
Cloning with `--single-branch` sets this config value to only fetch the
desired branch, so when we `git fetch origin <branch>` the new branch
is not available for `git checkout`. We can update this spec to point to
the new desired branch name instead.

We also make sure to `git fetch origin <branch>` before checking it out.
Otherwise git wouldn't know about the new branch yet.