Add a metadata variant for a reference to another parser in the repo
Some grammar directories in a repo might contain just queries if they
reuse another grammar's parser. We should reference this in the metadata
so that we can look up which grammar is used an assert that it is in
the same repository.

The idea is that the child commits will introduce functions to read
parser and query files from the skidder::Config. We should only load
query files if the parser is defined in the same repository to ensure
that we never load potentially mismatched queries. This is a problem
for languages like Starlark however which are defined as using another
language's parser. These directories currently don't have metadata
files since they don't have their own parsers.

We can add a variant of metadata that simply references another grammar
in the repository by name.

This also eliminates the need to configure this in a downstream
configuration like Helix's current `languages.toml` config. The
metadata instead is responsible for this relationship.