Fix layer finding comparisons
`partition_point` is always tricky. To find the injection that contains
the given point we should use `partition_point` to find the place where
the injections are start ending after the given point, then to find the
one injection that contains the point we need to check that the start
is less than or equal to the point. So `start <= point < end`.