Setup a test-grammars local repository for testing
Previously we had a dependency on the grammar repository that will be
used for language support in Helix. Instead this commit switches us to
a local repository inline here in the test-grammars directory.

We don't need Helix's full language support for testing and using that
repository means that we could need to update tests to accommodate for
updated grammars and queries in that repository. Instead we can
duplicate part of that repository here.

This also makes testing a bit simpler - previously we were downloading
a (currently private) repository when running `cargo test` which is a
bit icky and complicates the CI setup. (CI is currently failing because
we haven't given the proper credentials to the runner.) Testing with
locals files is a much smoother experience.