Locals highlighting
This builds on the prior prototype in fabc0f07. The `local.scope` and
`local.definition.*` captures are queried when an injection layer is
updated and the definitions are stored for future use in highlighting.
The highlight query also runs the part of the locals query which
captures `local.reference`. On that capture we lookup the reference
in the local scopes.

Per injection layer locals information is stored as a tree - scopes may
(and often do) overlap - which is similar in structure to the way that
injections are stored.