QueryIter: Replace Default bound with init function
Default is a reasonable bound for simplicity - we might also use it in
the function that exposes QueryIter in Helix - but it's limiting for the
highlighter. Per-layer we want a cursor over the local scopes in the
injection. The child commit will take advantage of this change to
initialize a locals cursor for the layer given these parameters.