Allow looking up by 'Match' injection marker in highlighter fixtures
This is used for example in testing Rust markdown codefences. It's a
simplistic example of how to use the 'Match' marker - typically you
would want more complicated rules like comparing the text to a regex -
but it should be enough for testing.

This change also refactors the InjectionLanguageMarker to be Copy
(which it is as of dccaa14 anyways) - this improves some ergonomics for
calling code, for example see how TestLanguageLoader implements the
callback that accepts this type.