highlighter: Add a QueryIter callback for checking predicates
This is hopefully extensible in the future. We need this for locals
because a locals pattern could contain multiple captures and we should
disable the pattern when a capture is not local. So this change moves
locals tracking into QueryIter and passes it in a new
`are_predicates_satisfied` callback.