highlighter: Represent `Highlight` as a non-max u32
`u32::MAX` is reserved for the null pointer space optimization:

    size_of::<Option<Highlight>>() == size_of::<Highlight>()

Unfortunately this kind of type is not available in std so we must roll
our own with bit-wise xor, which is an involution when used with a
constant like `u32::MAX`.

We could expose an interface like `NonZeroU32` and have `new` return an
`Option<Self>`. It is unlikely that a caller would ever pass `u32::MAX`
though so an assertion seems appropriate instead.