injections: Merge adjacent injections when reusing a layer
This fixes a case where layers would be re-parsed unnecessarily when
injections were reused. A minimal example is a markdown file like so:

    * Foo
    * Bar
      Baz

Editing "Foo" to add another 'o' for example causes the root layer and
the layer for "Fooo"'s line to be reparsed - correctly - but also causes
the other bullet item to be reparsed. This is because the other bullet
item is split into two injections in `intersect_ranges`. That causes the
layer to be re-parsed because `run_injection_query` considers the
layer to be modified because the range of the node matched with
`@injection.content` covers the combined range, not each range after
`intersect_ranges`. The solution here is to reconstruct the original
range by merging adjacent injections in `reuse_injection`.