bindings: Improve Debug output for Node
This resembles the debug impl from the `tree-sitter` crate - it's
showing the kind and range, though `tree-sitter` uses positions instead
of byte indices. This is clearer than the regular derived Debug impl
which shows the start byte (first entry in the context array) but not
anything else useful.

This also improves the Debug for Tree output since that type shows its
root node.