Remove path parameters from query initialization
The path was used for error reporting before. I feel this is better left
for the caller of these query initialization functions to include as
context. It's also not necessarily accurate because query text can be
concatenated from multiple sources, either because queries like
highlights and locals are merged together, or because of `; inherits`.