bindings: Eliminate function call overhead in `Node::start_byte`
We inline the implementation of `Node::start_byte` from `node.c` in the
C library. `Node::start_byte`'s presence is noticeable (although small)
in a flamegraph when highlighting in Helix. This change eliminates that
section of the flamegraph. It also improves a simple highlighting
benchmark noticeably (although not significantly: under 5%).

This is a tradeoff: this meaning of the entries in the `context` array
are not really part of the C library's public API. The implementation of
`ts_node_start_byte` is trivial though and, I expect, unlikely to change
in the future.