Use matched node byte range to detect modified reused layers
This commit reverts the change in eb5ca34 and replaces it with a more
robust mechanism. We need to detect when an injection, which is being
reused, is modified. The prior check compared the byte range of the
matched node (captured with `@injection.content`) to the range of the
injection, however the two are not the same thing. The range of the
injection is affected by `intersect_ranges` and can be trimmed at either
end or sliced up into many contained ranges. So instead of comparing the
injection range to the matched node range, we now store the byte range
of the node which was captured with `@injection.content` on `Injection`
and use that for comparison.