locals: Fix comparison conditions for creating the scope cursor
I updated this in be5c489 but didn't get the conditions quite right. We
want to skip over any children scopes that end before the given starting
position, so we should compare that the range end is less than the
position. The locals prototype in fabc0f0 was closer: it just needed the
sides of the comparison swapped.