highlighter: Avoid storing layer state in the QueryIter
The highlighter peeks the query iter so the query iter handles injection
events before the highlighter gets a chance to. Instead we can store
the current layer (before the peek) and the layer state on the
highlighter itself.

This fixes the tracking for the dormant and parent highlights -
previously if two injections started back-to-back we would mistakenly
put this information on the second injection (because the query iter had
been peeked).