highlighter: Fix partition_point condition for skipping injections
We want the QueryIter to skip any injections for a layer which come
before the requested query range. Previously the check allowed skipping
over an injection which started before the query range and ended after,
leading to missing highlights for the injected content. Instead we want
to skip past any injection ranges which end before the query start.