highlighter: Ensure injection ranges reside within parent layers
`intersect_ranges` should call `push_range` on ranges which reside both
within the node range (`range` parameter) and within parent ranges. In
an example Rust file like

    /// Hello world.
    fn main() {
        /// Hello, world!
        println("Hello, world!");
    }

editing the text to remove the leading '/' on line three would cause
`intersect_ranges` to be called with `range` corresponding to the
combined comment content (even though the comment on line three is no
longer a documentation comment) but with parent ranges covering only
the comment on line 1.

Within intersect_ranges, when the `parent_ranges` iterator runs dry, we
should return instead of break so that we don't attempt to call
`push_range` with a range that starts at the prior parent range start
and ends on `range.end`. This contains the ranges passed to `push_range`
to be within the parent ranges.