bindings: Revert switch to parser's parse_with_options API
My implementation appears to be unsound: when compiled with fat LTO,
Helix sees spurious failures to highlight claiming to be a timeout but
with a corrupt duration value. I'm reverting to the old
`ts_parser_set_timeout_micros` API since it avoids this issue for now.
With further debugging though we should switch back, especially before
the old API is dropped in v0.26.