highlighter: Avoid cloning injection properties
Cloning the properties can be avoided and it is relatively expensive in
terms of temporary allocations. The nice part of cloning was being able
to use `unwrap_or_default` to get default injection properties for
injections which don't use any `#set! injection.X` properties but we can
handle this where the properties are accessed instead.

Closes #12