highlighter: Consider layers finished if they have no more injections
Previously the `QueryIter` dropped layers which had no remaining
captures but the layer could potentially have more injections. In this
case the `QueryIter` would drop the layer and then recreate it later
when the parent layer entered another injection for this layer. This
could be potentially wasteful if the `QueryIter`'s `LayerState` type has
an expensive `Default` implementation.

We should prefer to emit `QueryIterEvent::ExitInjection` with `Some`
state (meaning that the layer is done and the state can be dropped) only
when the layer will not be entered again.