highlighter: Consider any skipped injection layers to be modified
This fixes an issue where an edit to a document which removes an
injection layer could eventually cause a panic that `parent_ranges` is
empty in `intersect_ranges`: `Option::unwrap` on a `None` value.

The test case edits a markdown block where a codefence is created
between two HTML comments. Before the edit, the document has a combined
injection for the two HTML comments. After the edit, the second HTML
comment is reinterpreted by the markdown parser as the content of the
codefence - not an HTML comment anymore.

In this case the HTML layer must be re-parsed because its included
ranges have been changed the edit: before it was a range for each
comment and after it is only the first comment. Previously we did not
mark the HTML layer as modified since it was not directly edited. But
this meant that, when running the HTML injection query over this layer
after the edit, we got query matches for the second HTML comment since
the node still existed according to the layer's tree. This second
injection capture did not intersect with any parent ranges since the
layer's ranges no longer included the second HTML comment, causing
`parent_ranges` to be empty and an unwrap of `None`.

The fix is to mark any layers which are not reused as modified. If a
layer is a combined injection and not all layers are reused then the
layer must be re-parsed because its included ranges have changed. All
modified layers are re-parsed. To do this we look at the old injections
for each layer and any that are skipped or discarded rather than reused
have their layer marked as modified.

Note that this does not cause layers which have been removed from the
tree to stick around. The layers are only marked as modified (meaning
that they need to be re-parsed) and not marked as touched (meaning that
they are still valid).

Also note that this is not an issue for adding injection ranges to a
combined injection since we always consider a layer to be modified if
we cannot reuse an injection.

A practical reproduction of this (and the source of the test) is
mentioned in helix-editor/helix#13544.