highlighter: Finish layers when all highlights on the layer are done
In the test cases the edoc layer (the backtick and single quote within
the comments) is recognized by `QueryIter` as being finished after it
highlights the inline code block - its only capture. The highlighter
needs to "deactivate" the layer rather than finishing (and removing it)
however, because the code block highlight extends to the next comment.

So when exiting an injection we check that the `QueryIter` considers a
layer to be finished and also that all highlights in the layer are done
as well.