Add NOTICE for erlang test grammar
Apache-2.0 on its own doesn't list the copyright holder since you are
not meant to edit the license text (unlike the MIT for example). So the
license file alone isn't enough, we also need the notice (recently
pushed upstream to tree-sitter-erlang). Also see the NOTICE in
tree-sitter-elixir which is also Apache-2.0 licensed.